material, hrmaterial entities deleted. Processtype converted to enum table

+++ src/HTS.Application.Contracts/Dto/ProcessCost/ProcessCostDto.cs
@@ -8,7 +8,7 @@ public class ProcessCostDto : EntityDto<int>
     public int ProcessId { get; set; }
     public DateTime ValidityStartDate { get; set; }
     public DateTime ValidityEndDate { get; set; }
-    public int HospitalPrice { get; set; }
-    public int UshasPrice { get; set; }
+    public decimal HospitalPrice { get; set; }
+    public decimal UshasPrice { get; set; }
     public bool IsActive { get; set; }
 }
